fix: do not duplicate Vite config
We already load and merge the vite.config.ts files from different roots,
but createServer from Vite (imported as createViteServer) will load the
the vite.config.ts from the project root and merge it with the provided
config. This means, in particular, that plugins may end up being
duplicated which can lead to various issues.