feat: allow to set title/level from preparser via frontmatter keys
I went with this solution as it doesn't change the `transformSlide` function signature. Otherwise, the change would either be breaking or we need to add the SlideInfo as redundant function argument (it also contains `content` and `frontmatter`). Having the whole SlideInfo might enable more use cases, but also might have the potential for modifications that shouldn't be done.

Closes #2095.